+++ src/components/BonusFlash/index.ts
@@ -1,0 +1,1 @@
+export { BonusFlash } from './BonusFlash'